% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bgindex.R
\name{compute_data_para}
\alias{compute_data_para}
\title{Internal helper: compute mixture parameters for a single gene}
\usage{
compute_data_para(exprs_mat_gene)
}
\description{
Internal helper: compute mixture parameters for a single gene
}
\keyword{internal}
