% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bgindex.R
\name{make_para_gn_parallel}
\alias{make_para_gn_parallel}
\title{Internal helper: parallel wrapper for compute_data_para}
\usage{
make_para_gn_parallel(exprs_mat, BPPARAM)
}
\description{
Internal helper: parallel wrapper for compute_data_para
}
\keyword{internal}
